chore: Run the same tests for different distributions
