--- cypress/integration/WithResizeEvent.spec.js
@@ -1,29 +1,0 @@
-describe('WithResizeEvent', () => {
-  beforeEach(() => {
-    cy.visit('/');
-    cy.get('textarea[is=enhanced-textarea]').as('textarea');
-    cy.window().then(win => win.defineElement());
-  });
-
-  afterEach(() => {
-    cy.wait(500);
-  });
-
-  it('Dispatches the userresize event', () => {
-    const handleUserResize = ({ target }) => {
-      target.setAttribute('resized', '');
-    };
-
-    cy.get('@textarea')
-      .invoke('on', 'userresize', handleUserResize)
-      .invoke('css', 'resize', 'both')
-      .wait(500)
-      .screenshot()
-      .trigger('pointerdown', 'bottomRight')
-      .invoke('css', 'height', '65px')
-      .trigger('pointerup')
-      .wait(500)
-      .screenshot()
-      .should('have.attr', 'resized');
-  });
-});
