+++ cypress/integration/builtIn.spec.js
@@ -1,0 +1,16 @@
+import runAutoHeightTests from '../tests/runAutoHeightTests.js';
+import runElementDefinitionTests from '../tests/runElementDefinitionTests.js';
+import runResizeEventTests from '../tests/runResizeEventTests.js';
+
+const selector = 'textarea[is=enhanced-textarea]';
+
+const options = {
+  selector,
+  startFunction: win => win.defineElement()
+};
+
+describe('Built-in', () => {
+  runAutoHeightTests(options);
+  runElementDefinitionTests(options);
+  runResizeEventTests(options);
+});
