+++ cypress/tests/runResizeEventTests.js
@@ -1,0 +1,33 @@
+const runResizeEventTests = ({ selector, startFunction }) => {
+  describe('Resize Event', () => {
+    beforeEach(() => {
+      cy.visit('/');
+      cy.get(selector).as('textarea');
+      cy.window().then(startFunction);
+    });
+
+    afterEach(() => {
+      cy.wait(500);
+    });
+
+    it('Dispatches the userresize event', () => {
+      const handleUserResize = ({ target }) => {
+        target.setAttribute('resized', '');
+      };
+
+      cy.get('@textarea')
+        .invoke('on', 'userresize', handleUserResize)
+        .invoke('css', 'resize', 'both')
+        .wait(500)
+        .screenshot()
+        .trigger('pointerdown', 'bottomRight')
+        .invoke('css', 'height', '65px')
+        .trigger('pointerup')
+        .wait(500)
+        .screenshot()
+        .should('have.attr', 'resized');
+    });
+  });
+};
+
+export default runResizeEventTests;
