+++ public/index.html
@@ -34,6 +34,7 @@
   <textarea
     rows="1"
     data="enhanced-textarea"
+    autoheight
   ></textarea>
 </body>
 </html>
